requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
torch==2.6.0
torchvision==0.21.0
numpy==2.2.4
pillow==11.1.0
opencv-python==4.11.0.86
timm==1.0.15
python-multipart==0.0.20
huggingface-hub==0.29.3
